chore: move ts-node closer to where it's used
